Create a new function for vimgrep, adjusted README
Created a new function that is able to do the following:

    - In Normal mode, search the current file using vimgrep for the word
      under the cursor.

    - In Visual mode, search the current file using vimgrep for the text
      that has been selected.

The replaces the old function, which only did the former.

In addition, the README file has been updated to include what will be a
new section containing a maintained list of all custom commands and
keyboard shortcuts.  Right now, this is just in comments, but will be
made live at some point.