Added new modules, new colors, changed vimrc
New colors include autumnleaf and summerfruit.  New modules include
ctrlp and syntastic.  Vimrc changes are mostly trivial, but include
increased functionality fo the FilterQuickFix function, which allows for
quickfix results to be filtered "against" a search pattern or file name.
The quickfix results will be filtered to *not* contain the specified
string or file name.